calendar-email-summary/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0
anthropic>=0.40.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
APScheduler>=3.10.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
